move ParsedPaymentOptions to own file
also smaller code  updates
