+++ src/lib/RequestTypes.ts
@@ -1,7 +1,5 @@
-type BigInteger = import('big-integer').BigInteger; // imports only the type without bundling
-import { FormattableNumber, toNonScientificNumberString } from '@nimiq/utils';
-import { isMilliseconds } from './Helpers';
-import { Currency, PaymentType, PaymentOptionsForCurrencyAndType, RequestType } from './PublicRequestTypes';
+import { Currency, PaymentType, RequestType } from './PublicRequestTypes';
+import { ParsedPaymentOptions } from './paymentOptions/ParsedPaymentOptions';
 import { ParsedNimiqSpecifics, ParsedNimiqDirectPaymentOptions } from './paymentOptions/NimiqPaymentOptions';
 import { ParsedEtherSpecifics, ParsedEtherDirectPaymentOptions } from './paymentOptions/EtherPaymentOptions';
 import { ParsedBitcoinSpecifics, ParsedBitcoinDirectPaymentOptions } from './paymentOptions/BitcoinPaymentOptions';
@@ -35,64 +33,6 @@ export type ParsedProtocolSpecificsForCurrency<C extends Currency> =
     : C extends Currency.BTC ? ParsedBitcoinSpecifics
     : C extends Currency.ETH ? ParsedEtherSpecifics
     : undefined;
-
-export interface ParsedPaymentOptions<C extends Currency, T extends PaymentType> {
-    readonly currency: C;
-    readonly type: T;
-    readonly decimals: number;
-    protocolSpecific: ParsedProtocolSpecificsForCurrency<C>;
-    amount: number | BigInteger;
-    expires?: number;
-    constructor: ParsedPaymentOptionsForCurrencyAndType<C, T>;
-    new(options: PaymentOptionsForCurrencyAndType<C, T>, ...optionalArgs: any[]):
-        ParsedPaymentOptionsForCurrencyAndType<C, T>;
-    raw(): PaymentOptionsForCurrencyAndType<C, T>;
-}
-
-export abstract class ParsedPaymentOptions<C extends Currency, T extends PaymentType>
-    implements ParsedPaymentOptions<C, T> {
-
-    protected constructor(options: PaymentOptionsForCurrencyAndType<C, T>) {
-        if (options.currency !== this.currency || options.type !== this.type) {
-            throw new Error(`Cannot parse given options as ${this.constructor.name}.`);
-        }
-        if (!this.isNonNegativeInteger(options.amount)) {
-            throw new Error('amount must be a non-negative integer');
-        }
-        this.expires = typeof options.expires === 'number'
-            ? isMilliseconds(options.expires)
-                ? options.expires
-                : options.expires * 1000
-            : undefined;
-    }
-
-    public get baseUnitAmount(): string {
-        return new FormattableNumber(this.amount).moveDecimalSeparator(-this.decimals).toString();
-    }
-
-    public update<P extends ParsedPaymentOptions<C, T>>(
-        this: P,
-        options: PaymentOptionsForCurrencyAndType<C, T>,
-        ...additionalArgs: any[]
-    ) {
-        const parsedOptions = new this.constructor(options as any, ...additionalArgs); // parse to check validity
-        this.amount = parsedOptions.amount; // amount must exist on all parsed options
-        this.expires = parsedOptions.expires || this.expires;
-        for (const key of
-            Object.keys(parsedOptions.protocolSpecific) as Array<keyof typeof parsedOptions.protocolSpecific>) {
-            if (parsedOptions.protocolSpecific[key] === undefined) continue;
-            this.protocolSpecific[key] = parsedOptions.protocolSpecific[key];
-        }
-    }
-
-    protected isNonNegativeInteger(value: string | number | bigint | BigInteger) {
-        try {
-            return /^\d+$/.test(toNonScientificNumberString(value));
-        } catch (e) {
-            return false;
-        }
-    }
-}
 
 export type AvailableParsedPaymentOptions = ParsedNimiqDirectPaymentOptions
                                           | ParsedEtherDirectPaymentOptions
